requirements.txt

reflex==0.8.15a1
ffmpeg-python
numpy
scipy
pillow
scipy
